startup_planner_backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
autopep8==2.3.1
awsebcli==3.20.10
blessed==1.20.0
botocore==1.31.85
cement==2.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-heroku==0.3.1
djangorestframework==3.15.2
duckduckgo_search==6.2.4
flake8==7.1.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
mccabe==0.7.0
openai==1.37.1
packaging==24.1
pathspec==0.10.1
primp==0.5.4
psycopg2==2.9.9
pycodestyle==2.12.0
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
resend==2.2.0
semantic-version==2.8.5
setuptools==70.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tabulate==0.9.0
tenacity==8.5.0
termcolor==1.1.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==1.26.19
vercel_blob==0.2.5
vercel_storage==0.0.1
wcwidth==0.1.9
wheel==0.43.0
whitenoise==6.7.0
